Merged PR 401457: Merge jserna/offsetstomping2 to future
Adding a usage sample for the new Translation Property availible in 15063. This sample shows how to use the new Translation property instead of Offset for a position animation. By avoiding the shared Offset property the animation works as designed. (This avoids the issue a.k.a Offset Stomping.) Using a simple keyframe animation on color rects defined in XAML the sample shows how to implement the new property. Includes a simple implicit animation that tilts the rects when window is resized as this type of animation is why Offset is shared.