Merged PR 232207: Merge ayoung/Interactions3DStuff to future
Removed the nested trackers and interaction sources.

We still don't want scale to affect position, so we're using Modifiers on the input (and inertia) center point, so that it always uses the world's top-left corner.

Moved Interactions3D to SDK Insider since we're using new RS2 APIs

Related work items: #10435809