Merged PR 1492495: 🐱‍👤 It turns out that when you make a color transparent, you can see what's behind it
Our splash screen/loading code has lots of issues.  This change doesn't fix that.  What this change does do is prevent the teaser effect on the main page where you see a peek of the fully loaded content before it animates in for realz.  The cause of this was simple - the fake splash screen we used has a color brush that's supposed to match the real splash screen's background and mask out all of the app content behind it.  Except that its alpha was set to 1.  Out of 255.  😿  While I was at it, I also tweaked the actual color being used to match what Visual Studio said the blue in the asset was.

- This doesn't fix our horrendous load time.  Empirically, this appears to be because of the telemetry shit that got pooped into our launch path.  Commenting out the telemetry events we fire on launch makes things much snappier.  I didn't look too deeply here, but I _believe_ setting the EndpointAddress in the AppTelemetryClient static constructor is what is particularly problematic.
- This doesn't fix the wonkiness in the splash screen code (extended splash screen anyone?)
- This new and improved experience does still have a cute effect of its own - you see the text on the real splash screen fade out before the fake one shows up to match it.  I *believe* this is because we're loading the surface for the fake splash screen asynchronously and it's just not there yet by the time the real one is going away.  This still looks way better than what we had before, and adding a synchronous path to the image loading code is a project for a different day.